Was recovered all code from old PR and made some changes
