+++ cpp/src/gandiva/function_holder_registry.h
@@ -25,10 +25,9 @@
 #include "arrow/status.h"
 #include "gandiva/function_holder.h"
 #include "gandiva/interval_holder.h"
-#include "gandiva/like_holder.h"
 #include "gandiva/node.h"
 #include "gandiva/random_generator_holder.h"
-#include "gandiva/replace_holder.h"
+#include "gandiva/regex_functions_holder.h"
 #include "gandiva/to_date_holder.h"
 
 namespace gandiva {
@@ -67,6 +66,7 @@ class FunctionHolderRegistry {
                                  {"random", LAMBDA_MAKER(RandomGeneratorHolder)},
                                  {"rand", LAMBDA_MAKER(RandomGeneratorHolder)},
                                  {"regexp_replace", LAMBDA_MAKER(ReplaceHolder)},
+                                 {"regexp_extract", LAMBDA_MAKER(ExtractHolder)},
                                  {"castintervalday", LAMBDA_MAKER(IntervalDaysHolder)},
                                  {"castintervalyear", LAMBDA_MAKER(IntervalYearsHolder)}};
     return maker_map;
